fix: the ability to delete someone else's location

+++ src/main/java/ru/weather/controllers/WeatherController.java
@@ -7,6 +7,7 @@ import org.springframework.web.bind.annotation.*;
 import ru.weather.dto.LocationDto;
 import ru.weather.dto.LocationJsonDto;
 import ru.weather.dto.WeatherDto;
+import ru.weather.models.Session;
 import ru.weather.models.User;
 import ru.weather.services.WeatherService;
 
@@ -18,8 +19,9 @@ public class WeatherController {
     private final WeatherService weatherService;
 
     @GetMapping("/")
-    public String mainPage(@RequestAttribute("user") User user,
+    public String mainPage(@RequestAttribute("session") Session session,
                            Model model) {
+        User user = session.getUser();
         List<WeatherDto> weather = weatherService.getWeatherByLogin(user.getLogin());
         model.addAttribute("weather", weather);
         model.addAttribute("user", user);
@@ -30,8 +32,9 @@ public class WeatherController {
 
     @GetMapping("/search-location")
     public String searchLocationByName(@RequestParam("name") String name,
-                                       @RequestAttribute("user") User user,
+                                       @RequestAttribute("session") Session session,
                                        Model model) {
+        User user = session.getUser();
         List<LocationJsonDto> locations = weatherService.searchLocationByName(name);
         model.addAttribute("user", user);
         model.addAttribute("locations", locations);
@@ -41,7 +44,8 @@ public class WeatherController {
 
     @PostMapping("/add-location")
     public String addLocation(@ModelAttribute LocationDto location,
-                              @RequestAttribute("user") User user) {
+                              @RequestAttribute("session") Session session) {
+        User user = session.getUser();
         location.setUser(user);
         weatherService.addLocation(location);
 
@@ -49,8 +53,12 @@ public class WeatherController {
     }
 
     @PostMapping("/delete-location")
-    public String deleteLocation(@RequestParam("id") Integer id) {
-        weatherService.deleteLocationById(id);
+    public String deleteLocation(@RequestParam("id") Integer id,
+                                 @RequestAttribute("session") Session session) {
+        if (weatherService.isLocationIdBelongsToUser(id, session.getUser())) {
+            weatherService.deleteLocationById(id);
+        }
+
         return "redirect:/";
     }
 }
